move login/register to its own component, redirect to player signup page upon registration
