+++ src/frontend/src/lib/components/NavBar.svelte
@@ -13,6 +13,7 @@
   import { ChevronDownOutline, ChevronDownSolid, BellSolid, BellOutline, GlobeSolid } from 'flowbite-svelte-icons';
   import AlertCount from './common/AlertCount.svelte';
   import { check_permission, series_permissions } from '$lib/util/permissions';
+  import LoginRegister from './common/LoginRegister.svelte';
   
   let notify: Notification;
 
@@ -88,8 +89,7 @@
           <ChevronDownSolid class="w-3 h-3 ms-2 text-white dark:text-white" />
         </Button>
         <Dropdown>
-          <DropdownItem href="/{$page.params.lang}/login">{$LL.NAVBAR.LOGIN()}</DropdownItem>
-          <DropdownItem href="/{$page.params.lang}/register">{$LL.NAVBAR.REGISTER()}</DropdownItem>
+          <LoginRegister/>
         </Dropdown>
       {/if}
       <NavHamburger/>
