+++ src/frontend/src/routes/[lang]/+page.svelte
@@ -6,6 +6,7 @@
   export let userId = 'Loading...';
   export let playerId = '';
   export let isLoggedIn = false;
+  import LoginRegister from '$lib/components/common/LoginRegister.svelte';
 
   onMount(async () => {
     const res = await fetch(`/api/user/me`);
@@ -20,27 +21,6 @@
       isLoggedIn = false;
     }
   });
-
-  async function loginOrSignup(event: SubmitEvent & { currentTarget: EventTarget & HTMLFormElement }) {
-    const data = new FormData(event.currentTarget);
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const payload = { email: data.get('email')!.toString(), password: data.get('password')!.toString() };
-
-    const isLogin = event.submitter?.classList.contains('login-btn') ?? false;
-    const endpoint = isLogin ? '/api/user/login' : '/api/user/signup';
-    console.log({ data, isLogin, endpoint });
-    const response = await fetch(endpoint, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(payload),
-    });
-
-    if (response.status < 300) {
-      goto('/');
-    } else {
-      alert(`${isLogin ? 'Login' : 'Registration'} failed`);
-    }
-  }
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   async function logout(event: SubmitEvent & { currentTarget: EventTarget & HTMLFormElement }) {
@@ -69,16 +49,5 @@
     <Button type="submit">{$LL.LOGOUT()}</Button>
   </form>
 {:else}
-  <form method="post" on:submit|preventDefault={loginOrSignup}>
-    <div>
-      <label for="email">{$LL.EMAIL()}</label>
-      <input name="email" type="email" />
-    </div>
-    <div>
-      <label for="password">{$LL.PASSWORD()}</label>
-      <input name="password" type="password" />
-    </div>
-    <button class="login-btn">{$LL.NAVBAR.LOGIN()}</button>
-    <button class="register-btn">{$LL.NAVBAR.REGISTER()}</button>
-  </form>
+  <LoginRegister/>
 {/if}
