--- src/frontend/src/routes/[lang]/login/+layout.svelte
@@ -1,5 +1,0 @@
-<svelte:head>
-  <title>Login | Mario Kart Central</title>
-</svelte:head>
-
-<slot />
