--- src/frontend/src/routes/[lang]/register/+layout.svelte
@@ -1,5 +1,0 @@
-<svelte:head>
-  <title>Register | Mario Kart Central</title>
-</svelte:head>
-
-<slot />
