+++ src/frontend/svelte.config.js
@@ -7,8 +7,6 @@ const locales = ['de', 'en-gb', 'en-us', 'es', 'fr', 'ja'];
 function getEntriesForLocale(locale) {
   return [
     `/${locale}`,
-    `/${locale}/login`,
-    `/${locale}/register`,
     `/${locale}/tournaments/details`,
     `/${locale}/tournaments/create`,
     `/${locale}/tournaments/create/select_template`,
